(feat) Finish implementing auto generation of react project structure (w/wo router).
