+++ server/structure/src/Index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Document</title>
+  <link rel="stylesheet" href="css/style.css">
+</head>
+<body>
+  <div id="App"></div>
+
+  <script src="./js/bundle.js"></script>
+</body>
+</html>
