+++ server/utils/worker.js
@@ -1,71 +1,23 @@
 const ejs = require('ejs');
 const fs = require('fs.extra');
 const path = require('path');
+const helper = require('./generateHelper');
 
-
-const webpackSetup = (tree, userId, cb) => {
-  const webpackTemplatePath = path.join(__dirname, '../templates/webpackConfigTemplate.ejs');
-  const webpackConfigPath = path.join(__dirname, `../../user/${userId}`);
-
-  ejs.renderFile(webpackTemplatePath, tree, (err, html) => {
-    fs.writeFile(`${webpackConfigPath}/webpack.config.js`, html, (err2) => {
-      if (err2) {
-        console.log(err2);
-      }
-      cb();
-    });
-  });
-};
-
-const htmlSetup = (tree, userId, callback) => {
-  const htmlTemplatePath = path.join(__dirname, '../templates/htmlTemplate.ejs');
-  const htmlPath = path.join(__dirname, `../../user/${userId}/src`);
-  const pageTemplatePath = path.join(__dirname, '../templates/page.ejs');
-  const pageJsonPath = path.join(__dirname, `../../user/${userId}`);
-
-  const page = tree.routes;
-  const pageObj = { pageData: [] };
-  const pageLength = tree.routes.length;
-  let counter = 0;
-
-  const generatePageJson = (cb) => {
-    page.forEach((single) => {
-      pageObj.pageData.push({
-        name: single.name,
-        path: `./src/js/${single.name}.jsx`,
-      });
-    });
-    ejs.renderFile(pageTemplatePath, pageObj, (err3, html) => {
-      fs.writeFile(`${pageJsonPath}/page.json`, html, (err4) => {
-        if (err4) {
-          console.log(err3);
-        }
-        cb();
-      });
-    });
-  };
-
-  const generateHtml = (treeData) => {
-    ejs.renderFile(htmlTemplatePath, treeData, (err, html) => {
-      fs.writeFile(`${htmlPath}/${treeData.name}.html`, html, (err2) => {
-        if (err2) {
-          console.log(err2);
-        }
-        counter += 1;
-        if (counter !== pageLength) {
-          generateHtml(page[counter]);
-        } else {
+const structureSetup = (tree, user, callback) => {
+  helper.webpackSetup(tree, user, () => {
+    helper.serverSetup(tree, user, () => {
+      if (tree.router === 1) {
+        helper.routerSetup(tree, user, () => {
           callback();
-        }
-      });
+        });
+      } else {
+        helper.htmlSetup(tree, user, () => {
+          callback();
+        });
+      }
     });
-  };
-
-  generatePageJson(() => {
-    generateHtml(page[counter]);
   });
 };
-
 
 module.exports = (tree, cb) => {
   const user = 1;
@@ -80,7 +32,7 @@ module.exports = (tree, cb) => {
 
   const generateFile = (treeData, inital) => {
     const tempTreeData = treeData;
-    if (inital) {
+    if (inital && tree.router === 0) {
       tempTreeData.initial = true;
     } else {
       tempTreeData.initial = false;
@@ -115,12 +67,10 @@ module.exports = (tree, cb) => {
         throw err2;
       }
       console.log("Copied 'structure' to 'user'");
-      htmlSetup(tree, user, () => {
-        webpackSetup(tree, user, () => {
-          for (let i = 0; i < tree.routes.length; i += 1) {
-            generateFile(tree.routes[i], true);
-          }
-        });
+      structureSetup(tree, user, () => {
+        for (let i = 0; i < tree.routes.length; i += 1) {
+          generateFile(tree.routes[i], true);
+        }
       });
     });
   });
